+++ src/main/java/camp/nextstep/edu/rebellion/domain/position/MovedPositions.java
@@ -1,0 +1,16 @@
+package camp.nextstep.edu.rebellion.domain.position;
+
+import java.util.Collections;
+import java.util.List;
+
+public class MovedPositions {
+    private final List<MovedPosition> positions;
+
+    public MovedPositions(List<MovedPosition> positions) {
+        this.positions = positions;
+    }
+
+    public List<MovedPosition> getPositions() {
+        return Collections.unmodifiableList(positions);
+    }
+}
