+++ src/main/java/camp/nextstep/edu/rebellion/game/origin/LadderLine.java
@@ -1,0 +1,63 @@
+package camp.nextstep.edu.rebellion.game.origin;
+
+import camp.nextstep.edu.rebellion.view.PrintableHorizonLineResolver;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static camp.nextstep.edu.rebellion.game.origin.LadderPointGenerator.generatePoint;
+
+public class LadderLine implements PrintableHorizonLineResolver {
+    private final List<Point> points;
+
+    public LadderLine(List<Point> points) {
+        this.points = points;
+    }
+
+    public int move(int position) {
+        return points.get(position).move();
+    }
+
+    public static LadderLine init(int sizeOfPerson) {
+        List<Point> points = new ArrayList<>();
+        Point point = initFirst(points);
+        point = initBody(sizeOfPerson, points, point);
+        initLast(points, point);
+        return new LadderLine(points);
+    }
+
+    public int getSizeOfPoint() {
+        return points.size();
+    }
+
+    private static Point initBody(int sizeOfPerson, List<Point> points, Point point) {
+        for (int i = 1; i < sizeOfPerson - 1; i++) {
+            point = point.next();
+            points.add(point);
+        }
+        return point;
+    }
+
+    private static void initLast(List<Point> points, Point point) {
+        point = point.last();
+        points.add(point);
+    }
+
+    private static Point initFirst(List<Point> points) {
+        Point point = Point.first(generatePoint());
+        points.add(point);
+        return point;
+    }
+
+    @Override
+    public String toString() {
+        return "LadderLine {" +
+                "points=" + points +
+                '}';
+    }
+
+    @Override
+    public boolean hasHorizonLine(int position) {
+        return position < move(position);
+    }
+}
