+++ src/main/java/camp/nextstep/edu/rebellion/game/origin/Point.java
@@ -1,0 +1,50 @@
+package camp.nextstep.edu.rebellion.game.origin;
+
+public class Point {
+    private final int index;
+    private final Direction direction;
+
+    public Point(int index, Direction direction) {
+        this.index = index;
+        this.direction = direction;
+    }
+
+    public int move() {
+        System.out.println("is left? " + direction.isLeft());
+        System.out.println("is right? " + direction.isRight());
+
+        if (direction.isRight()) {
+            return index + 1;
+        }
+
+        if (direction.isLeft()) {
+            return index - 1;
+        }
+
+        return this.index;
+    }
+
+    public Point next() {
+        return new Point(index + 1, direction.next());
+    }
+
+    public Point next(Boolean right) {
+        return new Point(index + 1, direction.next(right));
+    }
+
+    public Point last() {
+        return new Point(index + 1, direction.last());
+    }
+
+    public static Point first(Boolean right) {
+        return new Point(0, Direction.first(right));
+    }
+
+    @Override
+    public String toString() {
+        return "Point {" +
+                "index=" + index +
+                ", direction=" + direction +
+                '}';
+    }
+}
