+++ src/main/java/camp/nextstep/edu/rebellion/view/PrintableHorizonLineResolver.java
@@ -1,0 +1,6 @@
+package camp.nextstep.edu.rebellion.view;
+
+@FunctionalInterface
+public interface PrintableHorizonLineResolver {
+    boolean hasHorizonLine(int position);
+}
