--- src/test/java/camp/nextstep/edu/rebellion/refactoring/LadderLineTest.java
@@ -1,19 +1,0 @@
-package camp.nextstep.edu.rebellion.refactoring;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class LadderLineTest {
-    @Test
-    public void init() {
-        int sizeOfPerson = 5;
-        System.out.println(LadderLine.init(sizeOfPerson));
-    }
-
-    @Test
-    public void move() {
-        LadderLine line = LadderLine.init(2);
-        System.out.println("ladder result : " + line.move(0));
-    }
-}
